Consolidate imgproc.cc and imgproc_batch.cc into single file.
Created additional imgproc_util.[h|cc] files to save some code
that was no longer being used imgproc.cc, but will likely be
useful in the future.
