+++ src/imgproc/cpp/Makefile
@@ -1,13 +1,13 @@
 
 LIBS = -lopencv_core -lopencv_imgproc -lopencv_imgcodecs -lopencv_highgui
 
-all: imgproc_batch imgproc
+all: imgproc
 
-imgproc_batch: imgproc_batch.cc
-	g++ imgproc_batch.cc -o imgproc_batch $(LIBS)
+imgproc: imgproc.cc imgproc_util.o
+	g++ imgproc.cc imgproc_util.o -o imgproc $(LIBS)
 
-imgproc: imgproc.cc
-	g++ imgproc.cc -o imgproc $(LIBS)
+imgproc_util.o: imgproc_util.cc
+	g++ -c imgproc_util.cc -o imgproc_util.o $(LIBS)
 
 clean:
-	rm -f imgproc imgproc_batch
+	rm -f imgproc *.o
