+++ src/imgproc/cpp/imgproc_util.h
@@ -1,0 +1,27 @@
+/*
+   Copyright 2016 Google Inc.
+
+   Licensed under the Apache License, Version 2.0 (the "License");
+   you may not use this file except in compliance with the License.
+   You may obtain a copy of the License at
+
+   http://www.apache.org/licenses/LICENSE-2.0
+
+   Unless required by applicable law or agreed to in writing, software
+   distributed under the License is distributed on an "AS IS" BASIS,
+   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+   See the License for the specific language governing permissions and
+   limitations under the License.
+*/
+
+#include <opencv2/imgproc/imgproc.hpp>
+
+
+using namespace cv;
+
+
+float avgBWImagePixelValue(const Mat &image);
+
+float avgBWImagePixelValueInCircle(const Mat &image, const Vec3f &circle);
+
+void cropAroundSolarDisk(const Mat &src, Mat &dest, Vec3f disk);
